chore(nextjs-crud): merge util functions
