--- packages/nestjsx-crud/src/utils/generateSearchFilter.ts
@@ -1,10 +1,0 @@
-import { SCondition } from "@nestjsx/crud-request";
-import { CrudFilters } from "@refinedev/core";
-import { createSearchQuery } from "./createSearchQuery";
-
-export const generateSearchFilter = (filters: CrudFilters): SCondition => {
-    return createSearchQuery({
-        operator: "and",
-        value: filters,
-    });
-};
