+++ packages/nestjsx-crud/src/utils/index.ts
@@ -1,6 +1,3 @@
-export * from "./createSearchQuery";
-export * from "./generateSearchFilter";
-export * from "./generateSort";
 export * from "./handleFilter";
 export * from "./handleJoin";
 export * from "./handlePagination";
