+++ packages/nestjsx-crud/test/utils/handleFilter.spec.ts
@@ -1,6 +1,11 @@
-import { RequestQueryBuilder } from "@nestjsx/crud-request";
-import { CrudFilters, CrudSorting } from "@refinedev/core";
-import { handleFilter, handleSort } from "../../src/utils";
+import { RequestQueryBuilder, SCondition } from "@nestjsx/crud-request";
+import { CrudFilters, CrudSorting, CrudFilter } from "@refinedev/core";
+import {
+    handleFilter,
+    handleSort,
+    generateSearchFilter,
+    createSearchQuery,
+} from "../../src/utils";
 
 describe("handleFilter", () => {
     it("should apply filters to the query", () => {
@@ -65,5 +70,141 @@ describe("handleFilter", () => {
             "sort%5B0%5D=name%2CASC&s=%7B%22%24and%22%3A%5B%7B%22age%22%3A%7B%22%24gte%22%3A18%7D%7D%2C%7B%22email%22%3A%7B%22%24eq%22%3A%22john%22%7D%7D%5D%7D";
 
         expect(query.query()).toEqual(expectedQuery);
+    });
+});
+
+describe("generateSearchFilter", () => {
+    it("should generate a search filter with 'and' operator", () => {
+        const filters: CrudFilters = [
+            {
+                field: "name",
+                operator: "contains",
+                value: "John",
+            },
+            {
+                field: "age",
+                operator: "gte",
+                value: 25,
+            },
+        ];
+
+        const searchFilter = generateSearchFilter(filters);
+
+        expect(searchFilter).toEqual({
+            $and: [{ name: { $contL: "John" } }, { age: { $gte: 25 } }],
+        });
+    });
+
+    it("should generate an empty search filter with 'and' operator when no filters are provided", () => {
+        const filters: CrudFilters = [];
+
+        const searchFilter = generateSearchFilter(filters);
+
+        expect(searchFilter).toEqual({
+            $and: [],
+        });
+    });
+});
+
+describe("createSearchQuery", () => {
+    it("should create a search query for simple filters", () => {
+        const filter: CrudFilter = {
+            field: "testField",
+            operator: "gt",
+            value: 42,
+        };
+
+        const expectedQuery: SCondition = { testField: { $gt: 42 } };
+
+        expect(createSearchQuery(filter)).toEqual(expectedQuery);
+    });
+
+    it("should create a search query for complex filters", () => {
+        const filter: CrudFilter = {
+            operator: "and",
+            value: [
+                {
+                    field: "testField",
+                    operator: "lt",
+                    value: 10,
+                },
+                {
+                    field: "anotherField",
+                    operator: "ne",
+                    value: "testValue",
+                },
+            ],
+        };
+
+        const expectedQuery: SCondition = {
+            $and: [
+                {
+                    testField: {
+                        $lt: 10,
+                    },
+                },
+                {
+                    anotherField: {
+                        $ne: "testValue",
+                    },
+                },
+            ],
+        };
+
+        expect(createSearchQuery(filter)).toEqual(expectedQuery);
+    });
+
+    it("should create a search query for complex filters with nested filters", () => {
+        const filter: CrudFilter = {
+            operator: "and",
+            value: [
+                {
+                    field: "testField",
+                    operator: "lt",
+                    value: 10,
+                },
+                {
+                    operator: "or",
+                    value: [
+                        {
+                            field: "anotherField",
+                            operator: "ne",
+                            value: "testValue",
+                        },
+                        {
+                            field: "yetAnotherField",
+                            operator: "eq",
+                            value: "testValue2",
+                        },
+                    ],
+                },
+            ],
+        };
+
+        const expectedQuery: SCondition = {
+            $and: [
+                {
+                    testField: {
+                        $lt: 10,
+                    },
+                },
+                {
+                    $or: [
+                        {
+                            anotherField: {
+                                $ne: "testValue",
+                            },
+                        },
+                        {
+                            yetAnotherField: {
+                                $eq: "testValue2",
+                            },
+                        },
+                    ],
+                },
+            ],
+        };
+
+        expect(createSearchQuery(filter)).toEqual(expectedQuery);
     });
 });
